refactor(admin): Move `<Breadcrumbs>` component out of entities route
- We'd like to use the component later for the new `docs` route
  as well (not yet checked in).
- In the `docs` route, we want to use it with a different path prefix.
  Therefore, there is a new `pathPrefix` prop which we set to '/e'
  in the entities route.

Refs: TOCDEV-2560
