+++ packages/admin/src/components/Breadcrumbs/index.js
@@ -1,0 +1,1 @@
+export {default} from './Breadcrumbs'
