+++ packages/admin/src/routes/entities/components/Breadcrumbs/BreadcrumbsContainer.js
@@ -1,13 +1,14 @@
 import {connect} from 'react-redux'
 import {injectIntl} from 'react-intl'
 
-import Breadcrumbs from './Breadcrumbs'
+import Breadcrumbs from '../../../../components/Breadcrumbs'
 
 const mapActionCreators = {
 
 }
 
 const mapStateToProps = (state, props) => ({
+  pathPrefix: '/e',
   breadcrumbsInfo: state.entities.path.breadcrumbsInfo,
   currentViewTitle: state.entities.path.currentViewTitle
 })
